+++ dns_editor.php
@@ -14,9 +14,12 @@
 		$domain_id = intval($GLOBALS['tf']->variables->request['edit']);
 		$table = new TFTable;
 		$domain = get_dns_domain($domain_id);
+		$csrf_token = $table->csrf('dns_editor');
+		if (isset($GLOBALS['tf']->variables->request['update']))
+			$verify_csrf = verify_csrf('dns_editor');
 		if ($domain !== false)
 		{
-			if (isset($GLOBALS['tf']->variables->request['update']))
+			if (isset($GLOBALS['tf']->variables->request['update']) && $verify_csrf)
 			{
 				if (validate_input($GLOBALS['tf']->variables->request['update'], $domain_id, $GLOBALS['tf']->variables->request['type'], $GLOBALS['tf']->variables->request['content'], $GLOBALS['tf']->variables->
 					request['name'], $GLOBALS['tf']->variables->request['prio'], $GLOBALS['tf']->variables->request['ttl']))
@@ -52,7 +55,7 @@
 				unset($GLOBALS['tf']->variables->request['update']);
 				unset($GLOBALS['tf']->variables->request['record']);
 			}
-			if (isset($GLOBALS['tf']->variables->request['delete']) && $GLOBALS['tf']->variables->request['delete'] == 1)
+			if (isset($GLOBALS['tf']->variables->request['delete']) && $GLOBALS['tf']->variables->request['delete'] == 1 && $verify_csrf)
 			{
 				delete_dns_record($domain_id, $GLOBALS['tf']->variables->request['record']);
 				unset($GLOBALS['tf']->variables->request['delete']);
@@ -74,19 +77,14 @@
 					if (isset($GLOBALS['tf']->variables->request['record']) && $GLOBALS['tf']->variables->request['record'] == $record['id'])
 					{
 						$table->add_hidden('update', $record['id']);
-						$table->add_field("<table cellspacing=0 cellpadding=0><tr><td><input type=\"text\" name=\"name\" value=\"" . trim(str_replace($domain['name'], '', $record["name"]), '.') . "\" class=\"input\"></td><td>." .
-							$domain['name'] . "</td></tr></table>");
+						$table->add_field("<table cellspacing=0 cellpadding=0><tr><td><input type=\"text\" name=\"name\" value=\"" . trim(str_replace($domain['name'], '', $record["name"]), '.') . "\" class=\"input\"></td><td>." . $domain['name'] . "</td></tr></table>");
 						$sel = "<select name=\"type\">\n";
 						foreach (get_record_types() as $type_available)
 						{
 							if ($type_available == $record["type"])
-							{
 								$add = " SELECTED";
-							}
 							else
-							{
 								$add = "";
-							}
 							$sel .= " <option" . $add . " value=\"" . $type_available . "\" >" . $type_available . "</option>\n";
 						}
 						$sel .= "</select>\n";
@@ -102,26 +100,17 @@
 						$table->add_field($record['name']);
 						$table->add_field($record['type']);
 						if (strlen($record['content']) > 30)
-						{
 							$table->add_field('<a href="#" title="' . htmlspecial($record['content']) . '">' . substr($record['content'], 0, 30) . '...</a>');
-						}
 						else
-						{
 							$table->add_field($record['content']);
-						}
 						$table->add_field($record['ttl']);
 						if (in_array($record['type'], array('MX', 'SRV')))
-						{
 							$table->add_field($record['prio']);
-						}
 						else
-						{
 							$table->add_field();
-						}
 						//if ($record['type'] != 'SOA')
 						//{
-						$table->add_field($table->make_link('choice=none.dns_editor&edit=' . $domain_id . '&record=' . $record['id'], 'Edit') . ' ' . $table->make_link('choice=none.dns_editor&edit=' . $domain_id . '&record=' .
-							$record['id'] . '&delete=1', 'Delete'));
+						$table->add_field($table->make_link('choice=none.dns_editor&edit=' . $domain_id . '&record=' . $record['id'], 'Edit') . ' ' . $table->make_link('choice=none.dns_editor&edit=' . $domain_id . '&record=' . $record['id'] . '&delete=1&csrf_token=' . $csrf_token, 'Delete'));
 						//}
 						//else
 						//{
